Made web pages refresh on browser reload.
Made the beans @ViewScoped instead of @SessionScoped.
The beans don't really keep much state, so this should cause any extra
cost.

Data are loaded from the database and transferred directly to HTML and
the data in the bean is never reused, because data is reloaded from the
database the next time data is needed.